feat：流水线运营数据统计 #9735

+++ src/backend/ci/core/dispatch/biz-dispatch/src/main/kotlin/com/tencent/devops/dispatch/pojo/JobConcurrencyHistory.kt
@@ -31,7 +31,7 @@ import com.tencent.devops.dispatch.pojo.enums.JobQuotaVmType
 
 data class JobConcurrencyHistory(
     val projectId: String,
-    val jobConcurrency: Long,
+    val jobConcurrency: Int,
     val jobQuotaVmType: JobQuotaVmType,
     val createTime: String
 )
